+++ src/kernel/isr/irq_timer_c.c
@@ -6,5 +6,5 @@ void irq_timer_c(void) {
 }
 
 void irq_timer_register(void) {
-	idt_gate_set(0x20, (uint32_t)irq_timer_asm);
+	idt_gate_set(0x20, irq_timer_asm);
 }
